git-utils: split frozen status policy modules
